requirements.txt
PlexAPI
python-dotenv